cad glitch elimination approach

+++ inc/dataTypes.h
@@ -32,7 +32,7 @@ struct CAD_s
     uint8_t digit4;
     uint8_t digit5;
     uint8_t digit6;
-    uint16_t counter;
+    uint32_t counter;
     uint8_t updateStage;
     bool update;
 } cad;
